Move bannerTitle under banner folder
