+++ src/components/banner/bannerTitle/_spec.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { shallow } from 'enzyme'
-import theme from '../../theme/theme'
+import theme from '../../../theme/theme'
 import BannerTitle from './'
 import 'jest-styled-components'
