+++ src/components/banner/bannerTitle/index.js
