+++ src/components/bannerImage/_spec.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { shallow, mount } from 'enzyme'
 import toJSON from 'enzyme-to-json'
-import BannerTitle from '../bannerTitle'
+import BannerTitle from '../banner/bannerTitle'
 import BannerSubtitle from '../banner/bannerSubtitle'
 import BannerImage from '.'
 import 'jest-styled-components'
