+++ src/components/bannerImage/index.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import styled, { css } from 'styled-components'
 import { media } from '../../theme/media'
 import { Column, Row, Container } from '../grid'
-import BannerTitle from '../bannerTitle'
+import BannerTitle from '../banner/bannerTitle'
 import BannerSubtitle from '../banner/bannerSubtitle'
 import BannerDate from '../banner/bannerDate'
